Enable code analysis with TreatWarningsAsErrors and fix all violations
Turn on .NET code analysis via Directory.Build.props with AnalysisLevel
and EnforceCodeStyleInBuild. Configure rule severities in .globalconfig
with sensible defaults (bulk suggestion, important rules as warnings,
noisy/inapplicable rules suppressed).

Key fixes:
- Fix self-assignment bug in WaveOffsetStream (CA2245) with new
  RepositionSourceStream method, plus comprehensive unit tests
- Fix same pattern in MixDiffStream
- Add GC.SuppressFinalize to all Dispose methods (CA1816)
- Add base.Dispose(disposing) calls where missing (CA2215)
- Replace Exception base class with specific types (CA2201)
- Fix ArgumentException parameter ordering (CA2208)
- Make WaveCallbackInfo and WasapiOutRT implement IDisposable (CA1001)
- Fix IntPtr arithmetic to avoid overflow on 32-bit (CA2020)
- Add inexact Stream.Read check in WaveFileReader.GetChunkData (CA2022)

+++ NAudio.Wasapi/CoreAudioApi/AudioEndpointVolumeCallback.cs
@@ -56,7 +56,7 @@ namespace NAudio.CoreAudioApi
             //Determine offset in structure of the first float
             var offset = Marshal.OffsetOf<AudioVolumeNotificationDataStruct>("ChannelVolume");
             //Determine offset in memory of the first float
-            var firstFloatPtr = (IntPtr)((long)notifyData + (long)offset);
+            var firstFloatPtr = notifyData + (int)offset;
 
             var voldata = new float[data.nChannels];
